requirements.txt
Django
djangorestframework
django-cors-headers  # For handling Cross-Origin Resource Sharing (CORS)